refactor r_packet benchmark
